Added functionality for "merging" venues
Warning: Be aware that this may cause some issues
The way this works, is that the "aliases" field has been added to the Venue model.
This is a list containing names that we define to be names of venues equal to the one that the object represents.
Also, events are changed so that they refer to the name of the "mother"-venue when the original name is found to be just an alias.